[REFAC] Movie Entity `genreIds` 타입 변경
- `genreIds` 필드 및 관련 메서드 타입을 `List<Integer>` → `List<Long>`으로 수정
- 데이터 타입 명확화를 위한 개선

+++ src/main/java/com/insidemovie/backend/api/movie/entity/Movie.java
@@ -41,7 +41,7 @@ public class Movie {
     private String actors;              // 출연진
 
     @Column(name = "genre_ids")
-    private List<Integer> genreIds;     // 장르
+    private List<Long> genreIds;     // 장르
 
     private String title;                // 영화 제목
     private String titleEn;              // 영문 영화 제목
@@ -82,7 +82,7 @@ public class Movie {
         this.voteAverage = voteAverage;
     }
     // 장르 수정
-    public void updateGenreIds(List<Integer> genreIds) {
+    public void updateGenreIds(List<Long> genreIds) {
         this.genreIds = genreIds;
     }
     // 국가 수정
